refactor: remove realtime background logging
I've realized the iOS doesn't give the control to have the iPhone
dictate when the ring should log the health data and store that in
realtime. Instead we need to focus on having the ring take measurements
on its own and then get those historical logs whenever we sync. As such
I'm removing the timers and background tasks.